Shader redundancy issue fixed
Fixed a problem that happened when a Technique is nested into another
Technique: The sub-technique might have been already used to perform a
scene instanciation of other techniques (setupOverrides).
I this case, we might perform the instanciation many times on the same
technique if this technique is referenced >1 times in other techniques.
when this happens, I just ignore the redundance.

also removed the error message when the effect is not found.